Code improvements (#446) + reduced memory usage of RenderBucket by making "type" byte instead of int + improved code readability in TextItem by renaming variable + added latitude/longitude validation methods to GeoPointUtils + solved Java double precision problem in GeoPoint + made GeoPoint Serializable for easier Android usage + made BoundingBox.toString() human readable for better debugging

+++ vtm/src/org/oscim/renderer/bucket/LineBucket.java
@@ -81,7 +81,7 @@ public class LineBucket extends RenderBucket {
         this.level = layer;
     }
 
-    LineBucket(int type, boolean indexed, boolean quads) {
+    LineBucket(byte type, boolean indexed, boolean quads) {
         super(type, indexed, quads);
     }
